feat: POST /book-copies

+++ src/main/java/hei/school/demo/entity/BookCopy.java
@@ -1,5 +1,7 @@
 package hei.school.demo.entity;
 
+import java.math.BigDecimal;
+
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import hei.school.demo.entity.enums.BookFormat;
 import jakarta.persistence.*;
@@ -27,15 +29,15 @@ public class BookCopy {
   private BookFormat format;
 
   @Column(name = "selling_price")
-  private double sellingPrice;
-  //  @ManyToOne
-  //  @JoinColumn(name = "library_id", nullable = false)
-  //  private Library library;
+  private BigDecimal sellingPrice;
+  // @ManyToOne
+  // @JoinColumn(name = "library_id", nullable = false)
+  // private Library library;
 
-  //  @ManyToOne
-  //  @JoinColumn(name = "arrival_id")
-  //  private Arrival arrival;
-  //  @Column(name = "purchase_price", nullable = false)
-  //  private double purchasePrice;
-  //  private boolean sold;
+  // @ManyToOne
+  // @JoinColumn(name = "arrival_id")
+  // private Arrival arrival;
+  // @Column(name = "purchase_price", nullable = false)
+  // private double purchasePrice;
+  // private boolean sold;
 }
